[backend] fix: money & number float

+++ backend/source/php/model/EnumeratorValueModel.php
@@ -8,7 +8,7 @@ class EnumeratorValueModel extends Model
 {
     public ?int $id = null;
     public ?string $text = null;
-    public int|float|null $number = null;
+    public float|int|null $number = null;
     public ?string $date = null;
     public int $enumeratorId;
     public int $position;
